refactor dir name
